Installs package react navigation. Adds Navigation in App, creates new component AssignmentsIndex and sets it to first screen
