+++ App.js
@@ -1,18 +1,18 @@
 import React from "react";
-import { StyleSheet, Text, View } from "react-native";
+import { NavigationContainer } from "@react-navigation/native";
+import { createStackNavigator } from "@react-navigation/stack";
+import AssignmentsIndex from "./src/components/AssignmentsIndex";
+
+const Stack = createStackNavigator(); 
+
 const App = () => {
   return (
-    <View style={styles.container}>
-      <Text>Welcome to DevelUp</Text>
-    </View>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="develUp" component={AssignmentsIndex} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 };
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
+
 export default App;
